backend/requirements.txt
Django==5.0.0
django-ninja==1.1.0
django-cors-headers==4.3.1